Move SubsitutionTest to common package
because tested classes are also in common package
