+++ src/main/java/at/ac/tuwien/kr/alpha/grounder/NaiveGrounder.java
@@ -704,31 +704,6 @@ public class NaiveGrounder extends BridgedGrounder implements ProgramAnalyzingGr
 		throw new UnsupportedOperationException("Forgetting assignments is not implemented");
 	}
 
-	public static String groundAndPrintRule(NonGroundRule rule, Substitution substitution) {
-		StringBuilder ret = new StringBuilder();
-		if (!rule.isConstraint()) {
-			Atom groundHead = rule.getHeadAtom().substitute(substitution);
-			ret.append(groundHead.toString());
-		}
-		ret.append(" :- ");
-		boolean isFirst = true;
-		for (Atom bodyAtom : rule.getBodyAtomsPositive()) {
-			ret.append(groundLiteralToString(bodyAtom.toLiteral(), substitution, isFirst));
-			isFirst = false;
-		}
-		for (Atom bodyAtom : rule.getBodyAtomsNegative()) {
-			ret.append(groundLiteralToString(bodyAtom.toLiteral(false), substitution, isFirst));
-			isFirst = false;
-		}
-		ret.append(".");
-		return ret.toString();
-	}
-
-	static String groundLiteralToString(Literal literal, Substitution substitution, boolean isFirst) {
-		Literal groundLiteral = literal.substitute(substitution);
-		return  (isFirst ? "" : ", ") + groundLiteral.toString();
-	}
-
 	@Override
 	public NonGroundRule getNonGroundRule(Integer ruleId) {
 		return knownNonGroundRules.get(ruleId);
