+++ src/test/java/at/ac/tuwien/kr/alpha/common/SubstitutionTestUtil.java
@@ -1,4 +1,62 @@
+/**
+ * Copyright (c) 2016-2020, the Alpha Team.
+ * All rights reserved.
+ *
+ * Additional changes made by Siemens.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1) Redistributions of source code must retain the above copyright notice, this
+ *    list of conditions and the following disclaimer.
+ *
+ * 2) Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
 package at.ac.tuwien.kr.alpha.common;
 
+import at.ac.tuwien.kr.alpha.common.atoms.Atom;
+import at.ac.tuwien.kr.alpha.common.atoms.Literal;
+import at.ac.tuwien.kr.alpha.grounder.NonGroundRule;
+import at.ac.tuwien.kr.alpha.grounder.Substitution;
+
 public class SubstitutionTestUtil {
+
+	static String groundAndPrintRule(NonGroundRule rule, Substitution substitution) {
+		StringBuilder ret = new StringBuilder();
+		if (!rule.isConstraint()) {
+			Atom groundHead = rule.getHeadAtom().substitute(substitution);
+			ret.append(groundHead.toString());
+		}
+		ret.append(" :- ");
+		boolean isFirst = true;
+		for (Atom bodyAtom : rule.getBodyAtomsPositive()) {
+			ret.append(groundLiteralToString(bodyAtom.toLiteral(), substitution, isFirst));
+			isFirst = false;
+		}
+		for (Atom bodyAtom : rule.getBodyAtomsNegative()) {
+			ret.append(groundLiteralToString(bodyAtom.toLiteral(false), substitution, isFirst));
+			isFirst = false;
+		}
+		ret.append(".");
+		return ret.toString();
+	}
+
+	static String groundLiteralToString(Literal literal, Substitution substitution, boolean isFirst) {
+		Literal groundLiteral = literal.substitute(substitution);
+		return  (isFirst ? "" : ", ") + groundLiteral.toString();
+	}
 }
